+++ jaclang/__init__.py
@@ -5,7 +5,7 @@ import sys
 
 sys.path.insert(0, os.path.join(os.path.dirname(os.path.abspath(__file__)), "vendor"))
 
-from jaclang.plugin.builtin import JacBuiltin  # noqa: E402
+from jaclang.plugin.default import JacBuiltin  # noqa: E402
 from jaclang.plugin.default import JacFeatureDefaults  # noqa: E402
 from jaclang.plugin.feature import JacFeature, pm  # noqa: E402
 from jaclang.vendor import lark  # noqa: E402
